+++ lib/snowflex.ex
@@ -9,16 +9,15 @@ defmodule Snowflex do
   alias Ecto.Changeset
   alias Snowflex.Worker
 
-  @timeout :timer.seconds(60)
   @type query_param :: {:odbc.odbc_data_type(), list(:odbc.value())}
   @type sql_data :: list(%{optional(String.t()) => String.t()})
 
-  @spec sql_query(atom(), String.t(), non_neg_integer() | :infinity) ::
+  @spec sql_query(atom(), String.t(), timeout()) ::
           sql_data() | {:error, term}
-  def sql_query(pool_name, query, timeout \\ @timeout) do
+  def sql_query(pool_name, query, timeout) do
     case :poolboy.transaction(
            pool_name,
-           fn pid -> Worker.sql_query(pid, query, timeout) end,
+           &Worker.sql_query(&1, query, timeout),
            timeout
          ) do
       {:ok, results} -> process_results(results)
@@ -26,12 +25,12 @@ defmodule Snowflex do
     end
   end
 
-  @spec param_query(atom(), String.t(), list(query_param()), non_neg_integer() | :infinity) ::
+  @spec param_query(atom(), String.t(), list(query_param()), timeout()) ::
           sql_data() | {:error, term}
-  def param_query(pool_name, query, params \\ [], timeout \\ @timeout) do
+  def param_query(pool_name, query, params \\ [], timeout) do
     case :poolboy.transaction(
            pool_name,
-           fn pid -> Worker.param_query(pid, query, params, timeout) end,
+           &Worker.param_query(&1, query, params, timeout),
            timeout
          ) do
       {:ok, results} -> process_results(results)